Django-Boilerplate/requirements.prod.txt
gunicorn==20.1.0
psycopg2-binary==2.9.9
python-decouple==3.8
boto3==1.28.85
django-storages==1.14.2
django-cors-headers==4.3.1
whitenoise==6.6.0
